requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
bidict==0.22.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.5.0
fastapi==0.104.1
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.7
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
wsproto==1.2.0
